canonical URL, a11y, and LCP improvements
add metadata, moving splitting client and server component when needed as metadata is not allowed in a client component

reusable baseUrl

set x-url header for use in canonical urls for each page

remove hardcoded title tag

remove user-scalable=no to pass a11y checks, since it is ignored anyway by browsers now

LCP set as priority loading
